+++ .gitignore
@@ -14,3 +14,8 @@ Thumbs.db
 
 # Logs
 npm-debug.log
+
+package-lock.json
+
+.npm
+dist
