+++ index.ts
@@ -1,4 +1,5 @@
-import { httpServer } from "./src/http_server/index.js";
+import { httpServer } from "./src/http_server/index";
+import './src/ws_server';
 
 const HTTP_PORT = 8181;
